% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-constructors.R
\name{CartesianSquare}
\alias{CartesianSquare}
\title{Cartesian Square}
\usage{
CartesianSquare(x, ...)
}
\arguments{
\item{x}{A data frame or \code{\link[tibble]{tibble}}.}

\item{...}{Unquoted names of variables in \code{x}. These are first
\code{dplyr::select}ed from \code{x}, and only these variables are
square with \code{tidyr::crossing}.}
}
\value{
A \code{\link[tibble]{tibble}} whose rows denote the values of the
  Cartesian production of \code{x} with itself. If \code{x} has \code{n}
  columns, then the returned Cartesian square has \code{2n} columns: columns
  \code{1, ..., n} are the column names of \code{x} with \code{_i} appended;
  columns \code{n+1, ..., 2n} are the column names of \code{x} with
  \code{_j} appended.
}
\description{
Compute the Cartesian Square of a data frame or a tibble.
}
\details{
This is a thin wrapper for \code{\link[tidyr]{crossing}} that also performs
some string formatting on the column names of the data frame or tibble.
}
\seealso{
\code{\link[tidyr]{crossing}} for a description of how the rows of
  the Cartesian square are ordered.
}
