% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laplacian-eigenmaps.R
\name{D}
\alias{D}
\title{Degree Matrix}
\usage{
D(w, symmetric = FALSE)
}
\arguments{
\item{w}{A symmetric weighted adjacency matrix.}

\item{symmetric}{A logical. If \code{FALSE}, then \code{w} is checked for
symmetry. If it is known beforehand that \code{w} is symmetric, then it
is faster to call with \code{symmetric = TRUE}.}
}
\value{
A diagonal matrix, whose values on the main diagonal represent
  the weighted degrees of the vertices in \code{w}---i.e., the values along
  the main diagonal are the row-sums of \code{w}.
}
\description{
Compute the degree of each vertex in a weighted adjacency matrix.
}
