% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laplacian-eigenmaps.R
\name{L}
\alias{L}
\title{Laplacian Matrix}
\usage{
L(w, symmetric = FALSE)
}
\arguments{
\item{w}{A symmetric weighted adjacency matrix.}

\item{symmetric}{A logical. If \code{FALSE}, then \code{w} is checked for
symmetry. If it is known beforehand that \code{w} is symmetric, then it
is faster to call with \code{symmetric = TRUE}.}
}
\value{
A symmetric matrix, whose values on the main diagonal represent the
  degree of the vertices and whose values off the main diagonal represent
  the negative weights in the adjacency matrix.
}
\description{
Compute the Laplacian matrix from a weighted adjacency matrix.
}
\seealso{
\code{\link{D}}
}
