% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laplacian-eigenmaps.R
\name{LaplacianEigenmaps}
\alias{LaplacianEigenmaps}
\title{Laplacian Eigenmaps}
\usage{
LaplacianEigenmaps(w, symmetric = FALSE)
}
\arguments{
\item{w}{A symmetric weighted adjacency matrix.}

\item{symmetric}{A logical. If \code{FALSE}, then \code{w} is checked for
symmetry. If it is known beforehand that \code{w} is symmetric, then it
is faster to call with \code{symmetric = TRUE}.}
}
\value{
The eigenvalues \code{lambda} and corresponding eigenvectors \code{e}
  that are solutions to the generalized eigenvector problem:
  \code{L(w)*f = lambda*D(w)*e}.

  The eigenvalues and eigenvectors are returned as a list with two elements:
  \itemize{
    \item \code{values}: a numeric vector of the eigenvalues, in increasing
      order.
    \item \code{vectors}: a \code{\link[tibble]{tibble}} with columns
      \code{x, e1, ..., eN}, where \code{N} is the number of rows in the
      input matrix \code{w}. The column \code{x} contains the row names of
      \code{w}. The columns \code{e1, ..., eN} are the eigenvectors associated
      with the eigenvalues (e.g., \code{vectors$e1} is associated with
      \code{values[1]}). Columns \code{e2, ..., eJ} give an embedding, in
      \code{J-1}-dimensinal space, for the data identified by \code{x}.
  }
}
\description{
Compute the Laplacian-Eigenmaps embedding of a weighted adjacency matrix.
}
