% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-constructors.R
\name{Normalize}
\alias{Normalize}
\title{Normalize Weights}
\usage{
Normalize(x, weights = rlang::missing_arg())
}
\arguments{
\item{x}{A data frame or tibble.}

\item{weights}{The unquoted name of a numeric column in \code{x} whose values
will be shaped into a square adjacency matrix.}
}
\value{
A data frame that is like \code{x}, except that the values of the
  \code{weights} variable, \code{W(i,j)} have been normalized such that
  \code{W_tilde(i, j) = (i/n) * (W(i,j) / sqrt(E_x[W(i,x)] * E_x[W(j,x)]))}.
}
\description{
Normalize Weights
}
