% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phoneigen.R
\docType{package}
\name{phoneigen-package}
\alias{phoneigen-package}
\title{phoneigen: Phonetic Analyses with Laplacian Eigenmaps}
\description{
Functions for applying Laplacian Eigenmaps to acoustic data for
the purpose of phonetic analysis. Includes data sets of adult speakers'
productions of American English /s, S/ and of productions of /t, k/.
Vignettes apply Laplacian Eigenmaps to high-dimensional spectral
representations so as to learn low-dimensional representations for the
/s/-vs-/S/ and the /t/-vs-/k/ contrasts, respectively.
Accompanies the manuscript: A. R. Plummer and P. F. Reidy (submitted).
"Computing low-dimensional representations of speech from socio-auditory
structures for phonetic analyses". Journal of Phonetics.
}
