% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laplacian-eigenmaps.R
\name{ReduceDimensions}
\alias{ReduceDimensions}
\title{Reduce the Dimensionality of a Laplacian Eigenmaps Embedding}
\usage{
ReduceDimensions(x, ...)
}
\arguments{
\item{x}{The value of a call to \code{\link{LaplacianEigenmaps}}.}

\item{...}{Unquoted names of eigenvectors.}
}
\value{
A tibble.
}
\description{
Reduce a Laplacian Eigenmaps embedding to a small number of dimensions.
}
