% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-accessors.R
\name{StopBursts}
\alias{StopBursts}
\title{Stop Bursts}
\source{
See \url{learningtotalk.org} for more information about the
  Learning to Talk Project.
}
\usage{
StopBursts()
}
\value{
A \code{\link[tibble]{tibble}} with 1179 rows (i.e., stop productions)
  and 14 columns (i.e., variables):
  \itemize{
    \item \code{File}: The name of the WAV file, located in the directory
      \code{data-raw/wav-sibilants/} in the source package, that contains
      the uncompressed audio data for the production.
    \item \code{Participant}: An alphanumeric code for the participant who
      produced the production.
    \item \code{Session}: An alphanumeric code for the session during which
      the production was recorded. Each \code{Participant} code corresponds
      to 1 or 2 \code{Session} codes.
    \item \code{SessionCompleted}: The date and time that the participant
      completed the session, during which the production was recorded,
      in \code{YYYY-MM-DD HH:MM:SS} format.
    \item \code{TaggingCompleted}: The date and time that a research
      assistant completed tagging all productions from a given session,
      in \code{YYYY-MM-DD HH:MM:SS} format.
    \item \code{Trial}: The trial number within the session when the
      production was elicited.
    \item \code{Orthography}: The orthographic transcription of the word
      presented during the \code{Trial}, used to elicit a production of a
      sibilant fricative.
    \item \code{TargetC}: The WorldBet transcription of the target sibilant
      fricative.
    \item \code{TargetV}: The WorldBet trancscription of the vowel that
      immediately followed the target sibilant fricative.
    \item \code{TaggedBurst}: The time of the stop burst for the
      production in the original audio recording of the full session.
    \item \code{TaggedVOT}: The time of the onset of voicing for the
      production in the original audio recording of the full session.
    \item \code{Burst}: The time of the stop burst for the
      production in the WAV file listed in \code{File}. Always \code{0.02}.
    \item \code{VOT}: The time of the onset of voicing for the production
      in the WAV file listed in \code{File}.
    \item \code{ExcitationPattern}: A list-column of 361-component numeric
      vectors, each of which represents the values of an excitation pattern
      computed from the production. These values are associated to the
      vector of center frequencies, on the ERB scale:
      \code{seq(from = 3, to = 39, by = 0.1)}.
      See "Excitation Patterns" section below.
  }
}
\description{
Access the internal data set of stop bursts produced by 21 adult
native speakers of American English, who were recorded as part of the
Learning to Talk Project. The target productions were elicited in
word-initial, pre-vocalic position with a real word repetition task.
}
\section{Excitation Patterns}{

  An excitation pattern is a type of psychoacoustic spectrum that represents
  the distribution of auditory excitation across auditory filters. To compute
  an excitation pattern, the auditory periphery was modeled by a bank of 361
  bandpass filters. Each filter was a fourth-order, zero-phase gammatone filter.
  The center frequencies of the filters were uniformly spaced from 3 to 39
  along the ERB scale (i.e., 0.1 inter-filter spacing). The bandiwidth of
  each filter was proportional to its center frequency; hence, the filters
  were wider at high frequencies. These features model how the basilar
  membrane compresses the frequency scale logarithmically, and is
  differentially tuned to different frequency components. The excitation
  pattern of an acoustic waveform is computed by filtering it through the
  gammatone bank, summing the energy at the output of each filter, and
  associating the output energy of each filter to its center frequency.

  See \code{data-raw/03-excitation-patterns} in the source package for the
  code used to compute the vectors in the list-column \code{ExcitationPattern}
  of this data set.
}

\seealso{
\code{\link{StopBursts}}
}
