% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-constructors.R
\name{WeightEdgesIf}
\alias{WeightEdgesIf}
\title{Weight the Edges Between Nodes}
\usage{
WeightEdgesIf(x, weights = rlang::missing_arg(), condition, values)
}
\arguments{
\item{x}{A data frame or \code{\link[tibble]{tibble}}.}

\item{weights}{An unquoted name of the output variable where the weights are
populated.}

\item{condition}{An unquoted expression that, within \code{x}, evaluates to
a logical vector.}

\item{values}{An unquoted expression that, within \code{x}, evaluates to a
numeric vector. The values of the \code{name} variable when
\code{condition} evaluates to \code{TRUE}.}
}
\value{
A tibble that is like \code{x} except that it has a variable
  \code{name} whose values are determined by \code{condition} and
  \code{weights}.
}
\description{
Set the edge weights between nodes in a graph that is represented like
a \code{\link{CartesianSquare}} data frame or tibble.
}
