% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-constructors.R
\name{WeightEdgesIf}
\alias{WeightEdgesIf}
\title{Weight the Edges Between Nodes}
\usage{
WeightEdgesIf(x, edges, weights, name = rlang::missing_arg())
}
\arguments{
\item{x}{A data frame or \code{\link[tibble]{tibble}}.}

\item{edges}{An unquoted expression that, within \code{x}, evaluates to
a logical vector.}

\item{weights}{An unquoted expression that, within \code{x}, evaluates to a
numeric vector--the values of the \code{name} variable when
\code{edges} evaluates to \code{TRUE}.}

\item{name}{An unquoted name of the output variable where the weights are
populated.}
}
\value{
A tibble that is like \code{x} except that it has a variable
  \code{name} whose values are determined by \code{edges} and
  \code{weights}.
}
\description{
Set the edge weights between nodes in a graph that is represented like
a \code{\link{CartesianSquare}} data frame or tibble.
}
